добавить поддержку удаления карточки

+++ src/utils/api.js
@@ -67,12 +67,16 @@ class Api {
       .then(this._handleServerResponse)
     }
 
-    removeCard(cardId) {
-      return fetch(`${this._baseUrl}/cards/${cardId}`, {
-        method: 'DELETE',
-        headers: this._headers,
-      })
-      .then(this._handleServerResponse)
+    removeCard(cardId, isOwn) {
+      if(isOwn) {
+        return fetch(`${this._baseUrl}/cards/${cardId}`, {
+          method: 'DELETE',
+          headers: this._headers,
+        })
+        .then(this._handleServerResponse)
+      } else {
+        return console.log('Удалять карточки может только владелец карточки')
+      }
     }
 
     editProfileAvatar(avatar) {
